requirements.txt
asgiref==3.5.2
Django==4.0.5
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
psycopg2==2.9.3
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1

djangorestframework~=3.13.1
python-dotenv~=0.20.0